requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
crispy-bootstrap5==2024.2
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1.1
django-crispy-forms==2.3
djangorestframework==3.15.2
executing==2.1.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
gunicorn==23.0.0
h5py==3.11.0
idna==3.9
imbalanced-learn==0.12.3
imblearn==0.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.3
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
pyzmq==26.2.0
requests==2.32.3
rich==13.8.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
